requirements.txt
celery==4.2.1
Cython==0.26
dateutils==0.6.6
decorator==4.0.2
dj-database-url==0.5.0
Django==1.11.18
django-crispy-forms==1.7.2
django-filter==1.1.0
django-tables2==2.0.6
gnureadline==6.3.3
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2013.7
whitenoise==4.0
